fix: changed default enum tag from int32 to uint8

+++ crates/cyrusc_internal/src/cir/types.rs
@@ -107,7 +107,7 @@ impl CIREnumType {
                     None
                 }
             })
-            .unwrap_or_else(|| Box::new(CIRType::Plain(PlainType::Int32)))
+            .unwrap_or_else(|| Box::new(CIRType::Plain(PlainType::UInt8)))
     }
 
     #[inline]
